roles y permisos de la aplicacion 17 agosto 2021

+++ adsi20-app/resources/views/principal/users/partials/dataAction.blade.php
@@ -1,10 +1,17 @@
 <div class = "btn-group btn-group-sm">
 
-    <a href="{{ route('principal.users.show', $users->id) }}" class="btn btn-sm btn-secondary">Ver</a>
-    <a href="{{ route('principal.users.edit', $users->id) }}" class="btn btn-sm btn-success" style="margin-left: 10px !important">Editar</a>
+    @can('principal.users.show')
+        <a href="{{ route('principal.users.show', $users->id) }}" class="btn btn-sm btn-secondary">Ver</a>
+    @endcan
 
-    {{ Form::open(['route' => ['principal.users.destroy',$users->id], 'method' => 'DELETE'])}}
-        <button class="btn btn-sm btn-danger"style="margin-left: 10px !important">Eliminar</button>
-    {{ Form::close() }}
+    @can('principal.users.edit')
+        <a href="{{ route('principal.users.edit', $users->id) }}" class="btn btn-sm btn-success" style="margin-left: 10px !important">Editar</a>
+    @endcan
+
+    @can('principal.users.destroy')
+        {{ Form::open(['route' => ['principal.users.destroy',$users->id], 'method' => 'DELETE'])}}
+            <button class="btn btn-sm btn-danger"style="margin-left: 10px !important">Eliminar</button>
+        {{ Form::close() }}
+    @endcan
 
 </div>
